+++ AmethystAPI/src/minecraft/src-client/common/client/social/MultiplayerGameinfo.hpp
@@ -1,7 +1,6 @@
 #include <cstddef>
 #include <string>
 #include <vector>
-#include "GameConnectionInfo.hpp"
 #include <minecraft/src-deps/core/file/Path.hpp>
 
 namespace Social {
@@ -31,6 +30,17 @@ namespace Social {
         EDU = 0x0004,
         Mock = 0x0005,
         Invalid = 0xff,
+    };
+
+    class GameConnectionInfo {
+    public:
+        ConnectionType mType;
+        std::string mHostIpAddress;
+        std::byte padding1[4];
+        std::string mRakNetGUID;
+        uint32_t mHostPort;
+        std::byte padding2[20];
+        uint64_t mNetherNetID;
     };
 
     class MultiplayerGameinfo {
